+++ frontend/src/App.js
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from 'react';
 import { Provider } from 'react-redux';
 import './App.global.scss';
 import { TIMEOUT_MS } from './constants/FeatureToggles';
-import FeatureToggles from './FeatureToggles';
 import Routing from './router/Routing';
+import FeatureToggles from './utils/FeatureToggles';
 import store from './state';
 import { FEATURES_URL } from './constants/URL';
 import LoadingIndicator from './components/LoadingIndicator';
