+++ frontend/src/components/auth/RegisterForm.scss
