--- frontend/src/components/auth/SignUpForm.scss
@@ -1,12 +1,0 @@
-$color-page: #FFFFFF;
-$color-button: #252525;
-$color-text: #252525;
-$color-button-text: #FFFFFF;
-$color-primary: #7D357F;
-$color-box: #F5F5F5;
-$color-box-secondary: #DFDFDF;
-
-.box {
-    align-items: center;
-    margin-top: 10%;
-}
