+++ frontend/src/pages/Login.js
@@ -4,20 +4,19 @@ import { useDispatch, useSelector } from 'react-redux';
 import { useHistory, Redirect } from 'react-router-dom';
 import { login } from '../state/actions/auth';
 import Header from '../components/auth/Header';
-import SignInForm from '../components/auth/signInForm';
-import s from './SignIn.scss';
+import LoginForm from '../components/auth/LoginForm';
+import s from './Login.scss';
 import {
-  LOGGED_IN, LOGIN_FAILED, LOADING_LOGIN, LOADING_FETCH_SELF,
+  LOGGED_IN, LOGIN_FAILED, LOADING_LOGIN,
 } from '../constants/AuthStatus';
-import LoadingIndicator from '../components/LoadingIndicator';
 
-const SignIn = () => {
+const Login = () => {
   const auth = useSelector(state => state.auth);
   const { status } = auth;
 
   const dispatch = useDispatch();
 
-  const handleSignIn = user => dispatch(login(user));
+  const handleLogin = user => dispatch(login(user));
 
   if (status === LOGGED_IN)
     return <Redirect to="/home" />;
@@ -26,15 +25,15 @@ const SignIn = () => {
     alert('Incorrect email or password!');
 
   return (
-    <div className={s.signin}>
+    <div className={s.login}>
       <Layout cols={1}>
         <Header text="Login to get going" isLoading={status === LOADING_LOGIN} />
-        <SignInForm
-          onSignInUser={user => handleSignIn(user)}
+        <LoginForm
+          onLogin={user => handleLogin(user)}
         />
       </Layout>
     </div>
   );
 };
 
-export default SignIn;
+export default Login;
