+++ frontend/src/pages/Login.scss
@@ -1,0 +1,6 @@
+.login {
+    display: flex;
+    height: '1024px';
+    justify-content: space-around;
+    align-items: center;
+}
