+++ frontend/src/pages/Register.js
@@ -3,13 +3,12 @@ import { Layout } from 'wix-style-react';
 import { Redirect } from 'react-router';
 import { useSelector, useDispatch } from 'react-redux';
 import Header from '../components/auth/Header';
-import SignUpForm from '../components/auth/signUpForm';
-import s from './SignUp.scss';
+import RegisterForm from '../components/auth/RegisterForm';
+import s from './Register.scss';
 import { register, fetchSelf } from '../state/actions/auth';
 import { LOGGED_IN, REGISTER_SUCCEEDED } from '../constants/AuthStatus';
 
-// TODO: rename to Register
-const SignUp = () => {
+const Register = () => {
   const dispatch = useDispatch();
   const auth = useSelector(state => state.auth);
 
@@ -24,14 +23,14 @@ const SignUp = () => {
     return <Redirect to="/home" />;
 
   return (
-    <div className={s.signup}>
+    <div className={s.register}>
       <Layout cols={1}>
         <Header text="Finish your registration" />
-        <SignUpForm
+        <RegisterForm
           onRegister={user => registerUser(user)}
         />
       </Layout>
     </div>
   );
 };
-export default SignUp;
+export default Register;
