+++ frontend/src/pages/Register.scss
@@ -1,0 +1,6 @@
+.register {
+    display: flex;
+    height: '1024px';
+    justify-content: space-around;
+    align-items: center;
+}
