--- frontend/src/pages/SignIn.scss
@@ -1,15 +1,0 @@
-$color-page: #FFFFFF;
-$color-button: #252525;
-$color-text: #252525;
-$color-button-text: #FFFFFF;
-$color-primary: #7D357F;
-$color-box: #F5F5F5;
-$color-box-secondary: #DFDFDF;
-
-.signin {
-    background: $color-page;
-    display: flex;
-    height: '1024px';
-    justify-content: space-around;
-    align-items: center;
-}
