+++ frontend/src/router/GuestRoute.js
@@ -1,0 +1,1 @@
+// TODO: Route only for guests (non-authenticated users)
